Error messages are removed on action. Closes #80.
It was much harder than anticipated to make error messages fade over
time. The fading also adds unneeded attention to them and makes the
buttons in the user interface jump around, which is not what a stressed
meeting administrator needs, so instead the error messages are removed
when any button in the admin interface is clicked. Which makes buttons
move around when they are clicked instead... oh well, that's for another
issue and another commit.